requirements.txt
aiocontextvars==0.2.1
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
click==6.7
colorclass==2.2.0
colorlog==4.0.2
configparser==3.7.1
contextvars==2.3
croniter==0.3.27
emoji==0.5.1
emoji-extractor==1.0.17
findspark==1.3.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
idna-ssl==1.1.0
immutables==0.6
itsdangerous==1.1.0
Jinja2==2.10
kafka-python==1.4.4
MarkupSafe==1.1.0
mode==3.0.8
msgpack==0.6.1
multidict==4.5.2
mypy-extensions==0.4.1
nltk==3.4
py4j==0.10.4
pymongo==3.7.2
pyspark==2.1.2
pyspark-stubs==2.4.0.post1
python-dateutil==2.8.0
python-twitch-client==0.6.0
regex==2019.2.5
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.2.17
terminaltables==3.1.0
typing-extensions==3.7.2
urllib3==1.24.1
venusian==1.2.0
Werkzeug==0.14.1
yarl==1.3.0
